+++ Sources/ProjectDescription/Product.swift
@@ -11,6 +11,7 @@ public enum Product: String, Codable, Equatable {
     case unitTests = "unit_tests"
     case uiTests = "ui_tests"
     case bundle
+    case commandLineTool
     case appClip
 
     // Not supported yet